requirements.txt
streamlit
toml
python-dotenv
qdrant-client
sentence-transformers